Create setup for adding other visual display modes
