+++ osu.Game/Screens/Utility/LatencyVisualMode.cs
@@ -1,0 +1,13 @@
+// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+#nullable enable
+namespace osu.Game.Screens.Utility
+{
+    public enum LatencyVisualMode
+    {
+        Simple,
+        CircleGameplay,
+        ScrollingGameplay,
+    }
+}
