+++ osu.Game/Screens/Utility/ScrollingGameplay.cs
@@ -1,0 +1,12 @@
+// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+#nullable enable
+using osu.Framework.Graphics.Containers;
+
+namespace osu.Game.Screens.Utility
+{
+    public class ScrollingGameplay : CompositeDrawable
+    {
+    }
+}
